Use a gradient to show the grid.
This fixes #8.

Since it might be useful to test typecsset in old browsers, I have left
basehold.it as a fallback.

Modern browsers, however, will use gradients to show lines. This allows
using a gradient in an offline computer, for example (as long as that
computer is equipped with a modern browser, of course)